[FEATURE]: Refactor de package, creado y agregado interceptor, creado Dto de respuesta y usuario logueado. Creado recurso para login con json.
