+++ pom.xml
@@ -14,7 +14,7 @@
 	<name>loginAop</name>
 	<description>Proyecto base para loggin mediante aspecto.</description>
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>18</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
