+++ src/main/java/angema/base/loginAop/LogsAopApplication.java
@@ -1,0 +1,25 @@
+package angema.base.loginAop;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.client.RestTemplate;
+
+@CrossOrigin
+@EnableConfigurationProperties
+@SpringBootApplication
+public class LogsAopApplication {
+
+	@Bean
+	public RestTemplate restTemplate(RestTemplateBuilder builder) {
+		return builder.build();
+	}
+
+	public static void main(String[] args) {
+		SpringApplication.run(LogsAopApplication.class, args);
+	}
+
+}
