+++ src/main/java/angema/base/loginAop/app/registros/EmployeeRepository.java
@@ -1,0 +1,9 @@
+package angema.base.loginAop.app.registros;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EmployeeRepository extends JpaRepository < Employee, Long > {
+
+}
