+++ src/main/java/angema/base/loginAop/app/verificar/VerificarRequest.java
@@ -1,0 +1,10 @@
+package angema.base.loginAop.app.verificar;
+
+
+import lombok.Data;
+
+@Data
+public class VerificarRequest {
+    public String dni;
+    public String nroTramite;
+}
