+++ src/main/java/angema/base/loginAop/core/auth/AuthException.java
@@ -1,0 +1,15 @@
+package angema.base.loginAop.core.auth;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.UNAUTHORIZED)
+public class AuthException extends Exception {
+
+    private static final long serialVersionUID = 1L;
+
+    public AuthException(String message) {
+        super(message);
+    }
+
+}
