+++ src/main/java/angema/base/loginAop/core/auth/AuthRequest.java
@@ -1,0 +1,6 @@
+package angema.base.loginAop.core.auth;
+
+public class AuthRequest {
+    public String user;
+    public String password;
+}
