--- src/main/java/angema/base/logsAop/app/registros/EmployeeController.java
@@ -1,60 +1,0 @@
-package angema.base.logsAop.app.registros;
-
-import java.util.List;
-import java.util.Map;
-
-import angema.base.logsAop.core.exceptions.ResourceNotFoundException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.client.RestTemplate;
-
-
-@RestController
-@RequestMapping("/api/v1")
-public class EmployeeController {
-    @Autowired
-    private EmployeeService employeeService;
-    private static final String CREATE_EMPLOYEE_ENDPOINT_URL = "http://localhost:8080/api/v1/employees";
-    @GetMapping("/employees")
-    public List<Employee> getAllEmployees() {
-        Employee newEmployee = new Employee("admin", "admin", "admin@gmail.com");
-
-        RestTemplate restTemplate = new RestTemplate();
-        Employee result = restTemplate.postForObject(CREATE_EMPLOYEE_ENDPOINT_URL, newEmployee, Employee.class);
-        return employeeService.getAllEmployees();
-    }
-
-    @GetMapping("/employees/{id}")
-    public ResponseEntity<Employee> getEmployeeById(@PathVariable(value = "id") Long employeeId)
-            throws ResourceNotFoundException {
-        Employee employee = employeeService.getEmployeeById(employeeId)
-                .orElseThrow(() -> new ResourceNotFoundException("Employee not found for this id :: " + employeeId));
-        return ResponseEntity.ok().body(employee);
-    }
-
-    @PostMapping("/employees")
-    public Employee createEmployee(@RequestBody Employee employee) {
-        return employeeService.createEmployee(employee);
-    }
-
-    @PutMapping("/employees/{id}")
-    public ResponseEntity<Employee> updateEmployee(@PathVariable(value = "id") Long employeeId,
-                                                   @RequestBody Employee employeeDetails) throws ResourceNotFoundException {
-        Employee updatedEmployee = employeeService.updateEmployee(employeeId, employeeDetails);
-        return ResponseEntity.ok(updatedEmployee);
-    }
-
-    @DeleteMapping("/employees/{id}")
-    public Map<String, Boolean> deleteEmployee(@PathVariable(value = "id") Long employeeId)
-            throws ResourceNotFoundException {
-        return employeeService.deleteEmployee(employeeId);
-    }
-}
