--- src/main/java/angema/base/logsAop/app/registros/EmployeeRepository.java
@@ -1,9 +1,0 @@
-package angema.base.logsAop.app.registros;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface EmployeeRepository extends JpaRepository < Employee, Long > {
-
-}
