--- src/main/java/angema/base/logsAop/app/verificar/VerificarRequest.java
@@ -1,10 +1,0 @@
-package angema.base.logsAop.app.verificar;
-
-
-import lombok.Data;
-
-@Data
-public class VerificarRequest {
-    public String dni;
-    public String nroTramite;
-}
