--- src/main/java/angema/base/logsAop/core/auth/AuthResponse.java
@@ -1,23 +1,0 @@
-package angema.base.logsAop.core.auth;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class AuthResponse {
-
-    @JsonProperty("access_token")
-    private String accessToken;
-    @JsonProperty("token_type")
-    private String tokenType;
-    @JsonProperty("expires_in")
-    private Integer expiresIn;
-    @JsonProperty("refresh_token")
-    private String refreshToken;
-    @JsonProperty("issued_at")
-    private String issuedAt;
-    @JsonProperty("client_id")
-    private String clientId;
-}
