--- src/main/java/angema/base/logsAop/core/exceptions/ErrorDetails.java
@@ -1,28 +1,0 @@
-package angema.base.logsAop.core.exceptions;
-
-import java.util.Date;
-
-public class ErrorDetails {
-    private Date timestamp;
-    private String message;
-    private String details;
-
-    public ErrorDetails(Date timestamp, String message, String details) {
-        super();
-        this.timestamp = timestamp;
-        this.message = message;
-        this.details = details;
-    }
-
-    public Date getTimestamp() {
-        return timestamp;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public String getDetails() {
-        return details;
-    }
-}
