--- src/main/java/angema/base/logsAop/core/exceptions/ResourceNotFoundException.java
@@ -1,14 +1,0 @@
-package angema.base.logsAop.core.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends Exception{
-
-    private static final long serialVersionUID = 1L;
-
-    public ResourceNotFoundException(String message){
-        super(message);
-    }
-}
