+++ src/main/resources/application.properties
@@ -8,7 +8,7 @@ configs.auth.issuer=angema
 configs.auth.token.auth.path=/auth
 configs.auth.token.secret=secret
 configs.auth.token.expiration.seconds=3600
-configs.auth.exclude.paths="/auth,/health,/info"
+configs.auth.exclude.paths=/auth,/health,/info,/error
 
 server.error.include-stacktrace=never
