+++ src/test/java/angema/base/loginAop/LogsAopApplicationTests.java
@@ -1,0 +1,13 @@
+package angema.base.loginAop;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LogsAopApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
