--- src/test/java/angema/base/logsAop/LogsAopApplicationTests.java
@@ -1,13 +1,0 @@
-package angema.base.logsAop;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LogsAopApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
